% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicciones.R
\name{prediccion}
\alias{prediccion}
\title{Realiza una preddiccion}
\usage{
prediccion(Xn, ajuste)
}
\arguments{
\item{Xn}{(Datos con valores para predecir)}

\item{Ajuste}{(valores ajustados para hacer las predicciones)}
}
\value{
Lista de valores de prediccion.
}
\description{
Realiza una prediccion de los datos a partir del modelo ajustado.
}
