% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLM_Qual.R
\name{RLM_qual}
\alias{RLM_qual}
\title{Regresion Multiple Cualitativa}
\usage{
RLM_qual(data)
}
\arguments{
\item{data}{(archivo de datos)}

\item{Xn}{(valores para prediccion).}

\item{ajuste}{(modelo obtenido para prediccion).}
}
\value{
Lista con coeficientes de regresion, error estandar, r ajustado y r cuadrado.
}
\description{
Realiza una regresion lineal multiple con datos cualitativos.
}
