% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLM_Qual.R
\name{RLMqual}
\alias{RLMqual}
\title{Regresion Multiple Cualitativa}
\usage{
RLMqual(data)
}
\arguments{
\item{data}{(archivo de datos)}
}
\value{
Lista con coeficientes de regresion, error estandar, r ajustado y r cuadrado.
}
\description{
Realiza una regresion lineal multiple con datos cualitativos.
}
